增加了日志和管理系统的功能

+++ c++/myhead_file/event_manger/include/log/log.hpp
@@ -2,17 +2,21 @@
  * @file log.hpp
  * @author luoyebai (2112216825@qq.com)
  * @brief 日志系统的封装实现
- * @version 0.1
+ * @version 1.2
  * @date 2023-08-04
  *
  * @copyright Copyright (c) 2023
  *
  */
 
+// luoyebai TODO:fatal退出后保存栈堆
+
 #ifndef INCLUDE__LOG__HPP
 #define INCLUDE__LOG__HPP
 
 // std
+#include <sys/types.h>
+
 #include <cstdint>
 #include <cstdlib>
 #include <ctime>
@@ -23,7 +27,6 @@
 #include <mutex>
 #include <sstream>
 #include <string>
-#include <sys/types.h>
 // log sys
 #include "log_system.hpp"
 
@@ -33,9 +36,9 @@
  * 7类log宏,等级依次上升,Fatal会退出程序
  * dlog保证任何等级日志下都能看到数据通常不要使用
  */
-#define LOG(MODE, ...)                                                         \
-    log_system_ptr->addMsg(MODE, __PRETTY_FUNCTION__, __FILE__, __LINE__,      \
-                           ##__VA_ARGS__)
+#define LOG(MODE, ...)                                                    \
+    log_system_ptr->addMsg(MODE, __PRETTY_FUNCTION__, __FILE__, __LINE__, \
+                           std::this_thread::get_id(), ##__VA_ARGS__)
 #define LOG_UNKNOW(...) LOG(KUNKNOW, ##__VA_ARGS__)
 #define LOG_DEBUG(...) LOG(KDEBUG, ##__VA_ARGS__)
 #define LOG_INFO(...) LOG(KINFO, ##__VA_ARGS__)
@@ -73,8 +76,7 @@ void logSystemInit(const std::string &name, const std::string &dir_path = "../",
     std::array<std::string, 4> ymdh_str;
     for (size_t i = 0; i < 4; ++i) {
         ymdh_str[i] = std::to_string(ymdh[i]);
-        if (ymdh_str[i].length() == 1)
-            ymdh_str[i] = '0' + ymdh_str[i];
+        if (ymdh_str[i].length() == 1) ymdh_str[i] = '0' + ymdh_str[i];
     }
 
     // 日志路径初始化
@@ -84,24 +86,21 @@ void logSystemInit(const std::string &name, const std::string &dir_path = "../",
     // 删除过期日志
     if (fs::exists(log_dir_path)) {
         for (const auto &entry : fs::directory_iterator(log_dir_path)) {
-            if (!entry.is_directory())
-                continue;
+            if (!entry.is_directory()) continue;
             // 获取文件时间
             auto ftime =
                 std::chrono::file_clock::to_sys(fs::last_write_time(entry));
             // 当前时间
             const auto now = std::chrono::system_clock::now();
             auto diff_hours = (now - ftime) / 1h;
-            if (diff_hours > clear_timing)
-                fs::remove_all(entry);
+            if (diff_hours > clear_timing) fs::remove_all(entry);
         }
     }
 
     // 创建log目录
     log_dir_path += ymdh_str[0] + '-' + ymdh_str[1] + '-' + ymdh_str[2] + '-' +
                     ymdh_str[3] + '/';
-    if (!fs::exists(log_dir_path))
-        fs::create_directories(log_dir_path);
+    if (!fs::exists(log_dir_path)) fs::create_directories(log_dir_path);
 
     // 日志系统属性赋值
     log_system_ptr->author = name;
@@ -111,4 +110,4 @@ void logSystemInit(const std::string &name, const std::string &dir_path = "../",
     log_system_ptr->run();
 }
 
-#endif // !INCLUDE__LOG__HPP
+#endif  // !INCLUDE__LOG__HPP
